server/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\Projects\doxmytech\server\app\__init__.py: 1
# D:\Projects\doxmytech\server\app\facebook.py: 3
# D:\Projects\doxmytech\server\app\reddit.py: 4
# D:\Projects\doxmytech\server\app\stringToWordCloud.py: 3
# D:\Projects\doxmytech\server\app\twitter.py: 3,4
# D:\Projects\doxmytech\server\twitterWIP\dryruntweet.py: 1,2
Flask == 1.1.2

# D:\Projects\doxmytech\server\app\__init__.py: 2
Flask_Cors == 3.0.10

# D:\Projects\doxmytech\server\app\stringToWordCloud.py: 9
matplotlib == 3.3.3

# D:\Projects\doxmytech\server\app\facebook.py: 4
# D:\Projects\doxmytech\server\app\reddit.py: 5
# D:\Projects\doxmytech\server\app\stringToWordCloud.py: 4
# D:\Projects\doxmytech\server\app\twitter.py: 5
# D:\Projects\doxmytech\server\twitterWIP\dryruntweet.py: 3
requests == 2.24.0

# D:\Projects\doxmytech\server\app\stringToWordCloud.py: 6
wordcloud == 1.8.1

azure-ai-textanalytics == 5.0.0